Fix warningArrayList is a raw type. References to generic type ArrayList<E> should be parameterized
